+++ dev/test_async_queue.py
@@ -30,7 +30,6 @@ class BinanceDataProcessor:
         current_update_id = {}
         while self.ubwa.is_stop_request(stream_id=stream_id) is False:
             data = await self.ubwa.get_stream_data_from_asyncio_queue(stream_id)
-            b = data['lalaal']
             if data.get('data'):
                 market = str(data.get('stream').split('@')[0]).lower()
                 current_update_id[market] = data.get('data').get('lastUpdateId')
@@ -45,7 +44,6 @@ class BinanceDataProcessor:
         print(f"Start processing data of {stream_id} ...")
         while self.ubwa.is_stop_request(stream_id=stream_id) is False:
             data = await self.ubwa.get_stream_data_from_asyncio_queue(stream_id)
-            b = data['2222']
             # print(data)
             self.ubwa.asyncio_queue_task_done(stream_id)